# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Project & Task API",
        "description": "API RESTful para la gestión de proyectos y tareas",
        "contact": {
            "email": "ecuadra06@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Listar proyectos con filtros",
                "operationId": "fbc21feeb3df8a2cb519efb6480ae7d9",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filtrar por estado (active, inactive)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive"
                            ]
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nombre (parcial)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de proyectos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Project"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Crear un nuevo proyecto",
                "operationId": "a9d5bd5914cf75eaebafcf8b9e320dcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Proyecto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Obtener detalles de un proyecto",
                "operationId": "865085b82af27d647cf1f2a772aa7927",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del proyecto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Actualizar un proyecto existente",
                "operationId": "fe600b7ffce78456bce67c51b9eca51d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Proyecto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Project"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    },
                    "422": {
                        "description": "Validación fallida"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Eliminar un proyecto",
                "operationId": "d81ac2b0b51fa7143c69f1a5f4c27eb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID del proyecto",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Proyecto eliminado exitosamente"
                    },
                    "404": {
                        "description": "Proyecto no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Project": {
                "required": [
                    "id",
                    "name",
                    "status"
                ],
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "name": {
                        "type": "string",
                        "example": "Proyecto Alpha"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción detallada del proyecto.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-28T14:45:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-05-28T15:00:00Z"
                    }
                },
                "type": "object"
            },
            "ProjectCreateRequest": {
                "required": [
                    "name",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Proyecto Alpha"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción detallada del proyecto.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "active"
                    }
                },
                "type": "object"
            },
            "ProjectUpdateRequest": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Proyecto Beta Actualizado"
                    },
                    "description": {
                        "type": "string",
                        "example": "Descripción actualizada del proyecto.",
                        "nullable": true
                    },
                    "status": {
                        "type": "string",
                        "example": "inactive"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}